create make install target
Also refactor find dep apps script and add a generic moose_test_runner
script for installation.

Fixes #10786.
